+++ ft_strnstr.c
@@ -1,7 +1,9 @@
+#include "libft.h"
+
 char	*strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (!*needle)
